% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_files.R
\name{import_files}
\alias{import_files}
\title{Import QIIME2 artifacts and metadata}
\usage{
import_files(metadata="Path/metadata.tsv", table=NULL, table_norm=NULL, taxonomy=NULL)
}
\arguments{
\item{metadata}{(Required). A path to the Q2-formatted metadata.}

\item{table}{(Optional). A path to the Q2-feature-table artifact.}

\item{table_norm}{(Optional). A path to the normalized Q2-feature-table artifact.}

\item{taxonomy}{(Optional). A path to the Q2-taxonomy artifact for features in the provided feature table(s).}
}
\value{
A list object with the imported files.
}
\description{
Import QIIME2 artifacts and metadata
}
\examples{
library(freyR)
# Import QIIME2 artifacts and metadata
files <- import_files(metadata="Path/metadata.tsv",
                      table="Path/table_norm.qza",
                      table_norm="Path/table_norm.qza",
                      taxonomy="Path/taxonomy.qza")
}
