% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Calculate and plot alpha diversity results}
\usage{
plot_alpha(data=NULL, var1=NULL, var2=NULL, div=TRUE, idx="Shannon",
 facet.var2=FALSE, stats=TRUE, pairwise=TRUE, levels.1=NULL, levels.2=NULL, xlab=NULL, print=TRUE)
}
\arguments{
\item{data}{(Required). The list object imported with \code{\link{import_files}}.}

\item{var1}{(Required). Character string. variable to evaluate.}

\item{var2}{(Optional). Character string.
Secondary variable only for alpha diversity stats.}

\item{div}{(Required). Logical. Default \code{TRUE}.
Include diversity analysis with the index passed by the \code{idx} parameter.}

\item{idx}{(Required). Character string. Default \code{Shannon}.
If \code{div} is \code{TRUE}, it specified the diversity index to be used along with richness.}

\item{facet.var2}{(Required). Logical. Default \code{FALSE}.
Pass \code{var2} to \code{\link{facet_wrap}} in the box-plots.}

\item{stats}{(Required). Logical. Default \code{TRUE}.
Include overall Kruskal-Wallis results in the plot.}

\item{pairwise}{(Required). Logical. Default \code{TRUE}.
Include overall pairwise Wilcoxon results in the plot.
If \code{facet.var2} is \code{TRUE}, pairwise results will be represented by brackets.}

\item{levels.1}{(Optional). Character vector. Sort var1.}

\item{levels.2}{(Optional). Character vector. Sort var2.}

\item{xlab}{(Optional). Character string. Title for axis x in the plot.}

\item{print}{(Required). Logical. Default \code{TRUE}. Print the plot.}
}
\value{
Returns a list object and prints a plot.
}
\description{
Calculate and plot alpha diversity results
}
\examples{
library(freyR)
# Import QIIME2 artifacts and metadata
files <- import_files(metadata="Path/metadata.tsv", table_norm="Path/table_norm.qza", taxonomy="Path/taxonomy.qza")

# Calculate and plot alpha diversity results
alpha <- plot_alpha(files, var1="Treatment")
}
