% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta.R
\name{plot_beta}
\alias{plot_beta}
\title{Calculate and plot beta diversity results}
\usage{
plot_beta(data=NULL, var1=NULL, var2=NULL, weighted=FALSE, stats=TRUE,
 dist.u="jaccard", dist.w="jaccard", label=NULL, inter=FALSE, print=TRUE)
}
\arguments{
\item{data}{(Required). The list object imported with \code{\link{import_files}}.}

\item{var1}{(Required). Character string. Main variable to evaluate.
Stats results in the plot reference to this variable. This variable determines \code{color} in the plot.
If the number of groups within this variable exceeds the length of the palette,
a combination of colors and shapes will be used to represent the variable.
Consequently, it will not be possible to represent a second variable (\code{var2}) using shapes.}

\item{var2}{(Optional). Character string. This variable determines \code{shape} in the plot.}

\item{dist.u}{(Required). Character string. Default \code{"jaccard"}.
The dissimilarity index to be used in the binary version. See \code{\link{vegdist}} for more information.}

\item{weighted}{(Required). Logical. Default \code{FALSE}. Use weighted distance. This generates a second plot.}

\item{dist.w}{(Required). Character string. Default \code{"jaccard"}.
The dissimilarity index to be used in the non-binary version. See \code{\link{vegdist}} for more information.}

\item{label}{(Optional). Character string.
Use a third variable represented by \code{label} in the plot.}

\item{stats}{(Required). Logical. Default \code{TRUE}. Include PERMANOVA results in the plot.}

\item{inter}{(Required). Logical. Default \code{FALSE}.
If \code{var2} is not \code{NULL}, \code{inter} determines if the \code{\link{adonis2}} formula
will be set with interaction of factors (*) or not (+).}

\item{title}{(Optional). Character vector. Title for each plot.}

\item{print}{(Required). Logical. Default \code{TRUE}. Print the plot.}
}
\value{
Returns a list object and prints a plot.
}
\description{
Calculate and plot beta diversity results
}
\examples{
library(freyR)
# Import QIIME2 artifacts and metadata
files <- import_files(metadata="Path/metadata.tsv", table_norm="Path/table_norm.qza", taxonomy="Path/taxonomy.qza")

# Calculate and plot beta diversity results
beta <- plot_beta(files, var1="Treatment", var2="Location", title = c("uJaccard - Bacteria", "wJaccard - Bacteria"))
}
