% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa.R
\name{plot_taxa}
\alias{plot_taxa}
\title{Generates taxonomy bar plots and summarizes taxa relative abundances}
\usage{
plot_taxa(data=NULL, var1=NULL, var2=NULL, var3="Sample", type=c("Sample","Taxa"), taxrank=c("Species","Genus",...),
 cutoff=NULL, scales="free_x", top=FALSE, norm=c("rel", "log10", "raw"), levels.1=NULL, levels.2=NULL, nrow=1, print=TRUE)
}
\arguments{
\item{data}{(Required). The list object imported with \code{\link{import_files}}.}

\item{var1}{(Optional). Character string. Variable to be used by \code{\link{facet_wrap}}.
Typically, the larger groups of samples.}

\item{var2}{(Required). Character string. Samples will be grouped by this variable in the plot.
Each bar in the plot will represent the elements of this variable.
If you wish to represent individual samples, use \code{"Sample"}.
Use the same as \code{"taxrank"} to make each taxon is represented by a single bar
(e.g. \code{var2 = "Class", taxrank = "Class", type = "Taxa"}).}

\item{var3}{(Required). Character string. Default \code{"Sample"}. A third variable only for taxa summary.
Typically each individual sample.}

\item{type}{(Required). Character string. Default \code{"Sample"}. The plot type.
The currently supported options are \code{c("Sample", "Taxa")}.
Use the same as \code{"taxrank"} with \code{var2} to make each taxon is represented by a single bar
(e.g. \code{var2 = "Class", taxrank = "Class", type = "Taxa"}).}

\item{taxrank}{(Required). Character string. Default \code{"Class"}. Taxonomic level to represent in the plot.}

\item{cutoff}{(Optional). Numeric.
Determines which taxa will be grouped into ‘Others’ based on abundance (\code{top=FALSE}) or number of taxa (\code{top=TRUE}).}

\item{scales}{(Required). Character string. Default \code{"free_x"}. Customize axis scales.}

\item{top}{(Required). Logical. Default \code{FALSE}.
If \code{TRUE}, the \code{cutoff} parameter determines the absolute number of displayed taxa + 'Others'.}

\item{norm}{(Required). Character string. Default \code{"rel"} .The normalization method.
The currently supported options are \code{c("rel", "log10", "raw")}.}

\item{levels.1}{(Optional). Character vector. Sort var1.}

\item{levels.2}{(Optional). Character vector. Sort var2.}

\item{nrow}{(Required). Numeric. Default \code{1}.
If \code{facet} is \code{TRUE}, the number of rows to be used by \code{\link{facet_wrap}}.}

\item{print}{(Required). Logical. Default \code{TRUE}. Print the plot.}
}
\value{
Returns a list object and prints a plot.
}
\description{
Generates taxonomy bar plots and summarizes taxa relative abundances
}
\examples{
library(freyR)
# Import QIIME2 artifacts and metadata
files <- import_files(metadata="Path/metadata.tsv", table="Path/table.qza", taxonomy="Path/taxonomy.qza")

# Plot taxonomy
taxa <- plot_taxa(files, var1="Location", var2="Treatments", levels.1=c("A", "B", "C"), levels.2=c("A", "B", "C"))
}
