% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_table.R
\name{subsetSamples}
\alias{subsetSamples}
\title{Filter feature table based on metadata}
\usage{
subsetSamples(data=NULL, ..., norm=FALSE)
}
\arguments{
\item{data}{(Required). The list object imported with \code{\link{import_files}}.}

\item{...}{(Required). The sub-setting expression that can be passed to the function.
These arguments may be used to filter samples based on specific metadata.
For example, \code{Treatment == "Group A"} can be used to filter samples where the treatment is equal to "Group A".}

\item{norm}{(Required). Logical. Default \code{FALSE}.
If \code{TRUE}, filter normalized table.}
}
\value{
List with the filtered table.
}
\description{
Filter feature table based on metadata
}
\examples{
library(freyR)
# Import QIIME2 artifacts and metadata
files <- import_files(table="Path/table.qza", metadata="Path/metadata.tsv")

# Filter table based on metadata
table_filtered <- subsetSamples(files, Treatment == "Group A", norm = FALSE)
}
