% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_table.R
\name{subsetTaxa}
\alias{subsetTaxa}
\title{Filter feature table based on taxonomy}
\usage{
filter_samples(data=NULL, ..., freq = NULL, by=c("Taxa","Freq"), norm=FALSE)
}
\arguments{
\item{data}{(Required). The list object imported with \code{\link{import_files}}.}

\item{...}{(Optional). If \code{by = "Taxa"}, the sub-setting expression that can be passed to the function.
These arguments may be used to filter samples based on taxonomy.
For example, Genus == "Escherichia" can be used to filter features where the genus is equal to "Escherichia".}

\item{freq}{(Optional). If \code{by = "Freq"}, a function or a list of functions that take a vector of
abundance values and return a logical value. For example, function(x) sum(x > 1) > (0.3*length(x)) can be used
to filter features seen in at least 30 \% of the samples.}

\item{by}{(Required). Character string. Filter features based on taxonomy or frequency.
The currently supported options are \code{c("Taxa", "Freq")}.}

\item{norm}{(Required). Logical.
If \code{TRUE}, filter normalized table.}
}
\value{
List with the filtered table.
}
\description{
Filter feature table based on taxonomy
}
\examples{
library(freyR)
# Import QIIME2 artifacts and metadata
files <- import_files(table="Path/table.qza", taxonomy="Path/taxonomy.qza")

# Filter table based on taxonomy
table_filtered <- subsetTaxa(files, Genus == "Escherichia", by="Taxa", norm = FALSE)

# WARNING!
# When filtering out non-target taxa, this function will also remove any NA values by default.
# To include NA values in the subset, you need to modify your filter condition to explicitly allow NA values.
# This ensures that taxa with NA in the specified level are retained during the filtering process:
table_filtered <- filter_taxa(files_target, !Species == non_target | is.na(Species), by = "Taxa", norm = FALSE)

# Filter table based on frequency
#Remove taxa not seen more than 3 times in at least 20\% of the samples.
table_filtered <- subsetTaxa(files, freq = function(x) sum(x > 1) > (0.3*length(x)), by="Freq", norm = FALSE)
}
